Upload of Relocate-Sensorhubs.ps1
Relocates a Sensorhub to a new customer and OCC-Connector by modifying se3_cc.conf in the servereye directory.
First, the servereye services are stopped. Then the contents of se3_cc.conf are read and the lines customerID, parentGUID
and secretKey are replaced. Afterwards the services are started again, and depending on if the user passed the cleanup
argument, the folder ServerEye3 in ProgramData is deleted recursively.